all models running without error

+++ crossvalidate.py
@@ -25,61 +25,44 @@ import numpy as np
 
 
 def crossvalidate(Model,Data,Trainer,model_name,rmdirs,N):
-    choice =input("Cross-Validate(y/n)???")
-    if(choice =='y'):
-        print("its a yes!!\n Begin crossvaldations")
         
-        test_rates=np.logspace(-3,-10,num=6)
-        print("The following are the learning rates:",test_rates)
-        for eta in test_rates:
-            config = process_config('./configs/sf_optimized.json', model_name)
-            config.learning_rate=eta
-            #print("this is config:\n",config)
-            print("\nPresent learning rate:",eta)
-            
-            print("check config",config.learning_rate)
-            config.num_epochs=5
+    test_rates=np.logspace(-3,-6,num=6)
+    print("The following are the learning rates:",test_rates)
+    count=0
+    for eta in test_rates:
+        config = process_config('./configs/sf_optimized.json', model_name)
+        config.learning_rate=eta
+        #print("this is config:\n",config)
+        print("\nPresent learning rate:",eta)
+        
         #Remove summary dir, but not hyperparams               
-        
-            if rmdirs:
+        #config.num_epochs=5
+        if rmdirs:
+            try:
+                
+                #shutil.rmtree(config.summary_dir + '/test/')
+                #shutil.rmtree(config.summary_dir + '/train/')
+                shutil.rmtree(config.checkpoint_dir)
+            except (OSError, FileNotFoundError) :
+                print ('an error')
+                pass 
+        create_dirs([config.summary_dir, config.checkpoint_dir])
+        model = Model(config)
+        print("model:",model)
+        for idx in range(N):
+            print ('Running %s, iteration %d/%d'%(model_name, idx+1, N))
+            with tf.Session() as sess:
+                model.load(sess)
+                data = Data(config)
+                logger = Logger(sess, config)
+                trainer = Trainer(sess, model, data, config, logger)
+                #print("this is trainer:\n",trainer)
                 try:
-                    shutil.rmtree(config.summary_dir + '/test/')
-                    shutil.rmtree(config.summary_dir + '/train/')
-                    shutil.rmtree(config.checkpoint_dir)
-                except OSError:
-                    print ('an error')
-                    pass 
-            all_variables = tf.get_collection_ref(tf.GraphKeys.GLOBAL_VARIABLES)
-            model = Model(config)
-            print("model:",model)
-            for idx in range(N):
-                print ('Running %s, iteration %d/%d'%(model_name, idx+1, N))
-                with tf.Session() as sess:
-                    ########this part displays all the variables 
-                    sess.run(tf.variables_initializer(all_variables))
-                    print('------------------------------------------------------')
-                    for var in tf.global_variables():
-                        print('all variables: ' + var.op.name)
-                    for var in tf.trainable_variables():
-                        print('normal variable: ' + var.op.name)
-                    print('------------------------------------------------------')
-                    
-                    
-                    all_variables = tf.get_collection_ref(tf.GraphKeys.GLOBAL_VARIABLES)
-                    sess.run(tf.variables_initializer(all_variables))
-                    model.load(sess)
-                    data = Data(config)
-                    logger = Logger(sess, config)
-                    trainer = Trainer(sess, model, data, config, logger)
-                    #print("this is trainer:\n",trainer)
-                    try:
-                        trainer.train()
-                    except ValueError:
-                        print("Method fails to converge for these parameters")    
-                    
-    else:
-        print("Its a no(default is no)")
-        return
+                    trainer.train()
+                except ValueError:
+                    print("Method fails to converge for these parameters")   
+        count=count+1
+
 
     
 def main():
@@ -138,38 +121,44 @@ def main():
         Model = AEDFAModel
         Data = MNISTDataGenerator
         Trainer = AESFTrainer
-        
-    crossvalidate(Model,Data,Trainer,model_name,rmdirs,N)
-         #config = process_config('./configs/np_optimized.json', model_name)
-    config = process_config('./configs/sf_optimized.json', model_name)
-    #print("this is config:\n",config)
-    print("\n learning rate :",config.learning_rate)
-    #Remove summary dir, but not hyperparams      
-    if rmdirs:
-        try:
-            shutil.rmtree(config.summary_dir + '/test/')
-            shutil.rmtree(config.summary_dir + '/train/')
-            shutil.rmtree(config.checkpoint_dir)
-        except OSError:
-            print ('an error')
-            pass 
-
-    create_dirs([config.summary_dir, config.checkpoint_dir])
-
-    model = Model(config)
-    for idx in range(N):
-        print ('Running %s, iteration %d/%d'%(model_name, idx+1, N))
-        with tf.Session() as sess:	
-            #sess = tf.Session()
-            model.load(sess)
-            data = Data(config)
-            logger = Logger(sess, config)
-            trainer = Trainer(sess, model, data, config, logger)
-            
+    
+    choice =input("Cross-Validate(y/n)???")
+    if(choice =='y'):
+        print("its a yes!!\n Begin crossvaldations")
+        crossvalidate(Model,Data,Trainer,model_name,rmdirs,N)
+    
+    else:
+        print("Its a no(default is no)")
+        #config = process_config('./configs/np_optimized.json', model_name)
+        config = process_config('./configs/sf_optimized.json', model_name)
+        #print("this is config:\n",config)
+        print("\n learning rate :",config.learning_rate)
+        #Remove summary dir, but not hyperparams      
+        if rmdirs:
             try:
-                trainer.train()
-            except ValueError:
-                print("Method fails to converge for these parameters")
+                #shutil.rmtree(config.summary_dir + '/test/')
+                #shutil.rmtree(config.summary_dir + '/train/')
+                shutil.rmtree(config.checkpoint_dir)
+            except OSError:
+                print ('an error')
+                pass 
+
+        create_dirs([config.summary_dir, config.checkpoint_dir])
+
+        model = Model(config)
+        for idx in range(N):
+            print ('Running %s, iteration %d/%d'%(model_name, idx+1, N))
+            with tf.Session() as sess:	
+                #sess = tf.Session()
+                model.load(sess)
+                data = Data(config)
+                logger = Logger(sess, config)
+                trainer = Trainer(sess, model, data, config, logger)
+
+                try:
+                    trainer.train()
+                except ValueError:
+                    print("Method fails to converge for these parameters")
 
 if __name__ == '__main__':
     main()
